Refactors MainActivity.kt navigation logic to CoolTV.kt
